+++ runner/params.go
@@ -3,7 +3,7 @@ package runner
 import (
 	"time"
 
-	"github.com/DataDog/datadog-traceroute/common"
+	"github.com/DataDog/datadog-traceroute/traceroute"
 )
 
 type TracerouteParams struct {
@@ -14,7 +14,7 @@ type TracerouteParams struct {
 	MaxTTL                    int
 	Delay                     int
 	Timeout                   time.Duration
-	TCPMethod                 common.TCPMethod
+	TCPMethod                 traceroute.TCPMethod
 	DestinationPort           int
 	WantV6                    bool
 	TCPSynParisTracerouteMode bool
